pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = "Estrutura de aplicação RAG implementada para meu TCC"
authors = ["Marco Antônio <tchez.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
langchain-community = "^0.0.31"
faiss-cpu = "^1.8.0"
langchain-openai = "^0.1.1"
tiktoken = "^0.6.0"
langchain-text-splitters = "^0.0.1"
pinecone-client = "^3.2.2"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.3.5"
httpx = "^0.27.0"
isort = "^5.13.2"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff check . && blue --check . --diff && isort --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn rag.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=rag -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
